kyc db added

+++ app/src/main/java/debashis/me/ckycmanager/data/Dates.kt
@@ -1,3 +1,3 @@
 package debashis.me.ckycmanager.data
 
-data class Dates (val day:Byte,val month: Byte, val year: Byte)
+data class Dates (val day:Int,val month: Int, val year: Int)
